+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wagner/opmodes/mechanismtests/TestRegistrar.java
@@ -23,6 +23,7 @@ public final class TestRegistrar {
         tests.put("Claw", new TestClaw());
         tests.put("Arm", new TestArm());
         tests.put("RoadRunner", new TestRoadRunner());
+        tests.put("Motor Direction", new TestMotors());
 
         for(Map.Entry<String, LinearOpMode> set : tests.entrySet()) {
             OpModeMeta meta = new OpModeMeta.Builder()